iommu: Introduce struct iommu_context
This structure will hold a IOMMU context in a platform-anostic way.
It is meant to be used in place of domain_iommu for operations to happen in
specific IOMMU context rather than to whole Xen domain.

Allow devices to be associated to a context in addition to a domain to track
what devices are in which domains.

Signed-off-by Teddy Astie <teddy.astie@vates.tech>