iommu: Introduce and implement reattach_context
reattach_context is used to move a device from a context to another, changing the page table applied to the device.
This function is also used when destroying a context using IOMMU_TEARDOWNF_REATTACH_DEFAULT, in such case, the device is reattached to default context before destroying the context.

Signed-off-by Teddy Astie <teddy.astie@vates.tech>