iommu: Introduce context parameter to IOMMU operations
Allow specifying on what IOMMU context of a Xen domain to apply the operation,
effectively allowing usage of several contexts within a same Xen domain.

Using 0 as ctx_no is same as previous behavior.

arch_iommu_context is meant to be the platform-specific representation of a IOMMU context.

For x86, this stores the address of the page table and the pages that are used
to build it. It is meant to be used to split arch_iommu context-specific parts
in a specific structure.

Signed-off-by Teddy Astie <teddy.astie@vates.tech>